# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseFields = {[key: string]: string};

export interface ErrorResponse {
  error: string;
  fields?: ErrorResponseFields;
}

export type AvailabilityLevel = typeof AvailabilityLevel[keyof typeof AvailabilityLevel];


export const AvailabilityLevel = {
  available: 'available',
  limited: 'limited',
  full: 'full',
} as const;

export interface AvailabilityResponse {
  level: AvailabilityLevel;
  remainingSeats: number;
  capacity: number;
}

export interface ReservationRequest {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  email: string;
  /**
     * @minLength 7
     * @maxLength 20
     */
  phone: string;
  date: string;
  /** @pattern ^([01]\d|2[0-3]):[0-5]\d$ */
  time: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  guests: number;
  /** @maxLength 500 */
  dietaryNotes?: string;
}

export interface ReservationResponse {
  id: string;
  confirmationCode: string;
  name: string;
  email: string;
  phone: string;
  date: string;
  time: string;
  guests: number;
  dietaryNotes?: string | null;
  depositRequired: boolean;
  depositAmount: number;
  status: string;
  createdAt: string;
}

export type GiftCardRequestRecipientType = typeof GiftCardRequestRecipientType[keyof typeof GiftCardRequestRecipientType];


export const GiftCardRequestRecipientType = {
  myself: 'myself',
  gift: 'gift',
} as const;

export type GiftCardRequestDeliveryMethod = typeof GiftCardRequestDeliveryMethod[keyof typeof GiftCardRequestDeliveryMethod];


export const GiftCardRequestDeliveryMethod = {
  email: 'email',
  physical: 'physical',
} as const;

export interface GiftCardRequest {
  /**
     * @minimum 500
     * @maximum 50000
     */
  amount: number;
  recipientType: GiftCardRequestRecipientType;
  deliveryMethod: GiftCardRequestDeliveryMethod;
  /** @maxLength 120 */
  recipientName?: string;
  recipientEmail?: string;
  /** @maxLength 500 */
  mailingAddress?: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  senderName: string;
  senderEmail: string;
  /** @maxLength 200 */
  message?: string;
}

export interface GiftCardResponse {
  id: string;
  code: string;
  amount: number;
  recipientType: string;
  deliveryMethod: string;
  status: string;
  createdAt: string;
}

export type ContactRequestSubject = typeof ContactRequestSubject[keyof typeof ContactRequestSubject];


export const ContactRequestSubject = {
  general: 'general',
  private: 'private',
  press: 'press',
  careers: 'careers',
} as const;

export interface ContactRequest {
  subject: ContactRequestSubject;
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  email: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
}

export interface ContactResponse {
  id: string;
  status: string;
  createdAt: string;
}

/**
 * The request failed validation
 */
export type ValidationErrorResponse = ErrorResponse;

export type GetAvailabilityParams = {
date: string;
/**
 * @pattern ^([01]\d|2[0-3]):[0-5]\d$
 */
time: string;
/**
 * @minimum 1
 * @maximum 20
 */
guests: number;
};