# packages/api-zod/src/generated/types/availabilityResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AvailabilityLevel } from './availabilityLevel';

export interface AvailabilityResponse {
  level: AvailabilityLevel;
  remainingSeats: number;
  capacity: number;
}