# packages/api-zod/src/generated/types/contactResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactResponse {
  id: string;
  status: string;
  createdAt: string;
}