# packages/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ErrorResponseFields } from './errorResponseFields';

export interface ErrorResponse {
  error: string;
  fields?: ErrorResponseFields;
}