# packages/api-zod/src/generated/types/giftCardResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GiftCardResponse {
  id: string;
  code: string;
  amount: number;
  recipientType: string;
  deliveryMethod: string;
  status: string;
  createdAt: string;
}