# packages/api-zod/src/generated/types/reservationResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationResponse {
  id: string;
  confirmationCode: string;
  name: string;
  email: string;
  phone: string;
  date: string;
  time: string;
  guests: number;
  dietaryNotes?: string | null;
  depositRequired: boolean;
  depositAmount: number;
  status: string;
  createdAt: string;
}